+++ evm/src/test/java/org/hyperledger/besu/evm/operation/LondonSStoreOperationGasCostTest.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.evm.operations;
+package org.hyperledger.besu.evm.operation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,46 +23,50 @@ import org.hyperledger.besu.evm.internal.EvmConfiguration;
 import org.hyperledger.besu.evm.testutils.TestCodeExecutor;
 
 import org.apache.tuweni.units.bigints.UInt256;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class ConstantinopleSStoreOperationGasCostTest {
+public class LondonSStoreOperationGasCostTest {
 
   public static Object[][] scenarios() {
-    // Tests specified in EIP-1283.
+    // Tests specified in EIP-3539.
     return new Object[][] {
-      {"0x60006000556000600055", 0, 412, 0},
-      {"0x60006000556001600055", 0, 20212, 0},
-      {"0x60016000556000600055", 0, 20212, 19800},
-      {"0x60016000556002600055", 0, 20212, 0},
-      {"0x60016000556001600055", 0, 20212, 0},
-      {"0x60006000556000600055", 1, 5212, 15000},
-      {"0x60006000556001600055", 1, 5212, 4800},
-      {"0x60006000556002600055", 1, 5212, 0},
-      {"0x60026000556003600055", 1, 5212, 0},
-      {"0x60026000556001600055", 1, 5212, 4800},
-      {"0x60026000556002600055", 1, 5212, 0},
-      {"0x60016000556000600055", 1, 5212, 15000},
-      {"0x60016000556002600055", 1, 5212, 0},
-      {"0x60016000556001600055", 1, 412, 0},
-      {"0x600160005560006000556001600055", 0, 40218, 19800},
-      {"0x600060005560016000556000600055", 1, 10218, 19800},
-      {"0x60026000556000600055", 1, 5212, 15000},
+      {"0x60006000556000600055", 0, 212, 0},
+      {"0x60006000556001600055", 0, 20112, 0},
+      {"0x60016000556000600055", 0, 20112, 19900},
+      {"0x60016000556002600055", 0, 20112, 0},
+      {"0x60016000556001600055", 0, 20112, 0},
+      {"0x60006000556000600055", 1, 3012, 4800},
+      {"0x60006000556001600055", 1, 3012, 2800},
+      {"0x60006000556002600055", 1, 3012, 0},
+      {"0x60026000556000600055", 1, 3012, 4800},
+      {"0x60026000556003600055", 1, 3012, 0},
+      {"0x60026000556001600055", 1, 3012, 2800},
+      {"0x60026000556002600055", 1, 3012, 0},
+      {"0x60016000556000600055", 1, 3012, 4800},
+      {"0x60016000556002600055", 1, 3012, 0},
+      {"0x60016000556001600055", 1, 212, 0},
+      {"0x600160005560006000556001600055", 0, 40118, 19900},
+      {"0x600060005560016000556000600055", 1, 5918, 7600},
     };
+  }
+
+  private TestCodeExecutor codeExecutor;
+
+  @BeforeEach
+  public void setUp() {
+    codeExecutor = new TestCodeExecutor(MainnetEVMs.london(EvmConfiguration.DEFAULT));
   }
 
   @ParameterizedTest
   @MethodSource("scenarios")
-  void shouldCalculateGasAccordingToEip1283(
+  void shouldCalculateGasAccordingToEip3529(
       final String code,
       final int originalValue,
       final int expectedGasUsed,
       final int expectedGasRefund) {
-
-    TestCodeExecutor codeExecutor =
-        new TestCodeExecutor(MainnetEVMs.constantinople(EvmConfiguration.DEFAULT));
-
     final long gasLimit = 1_000_000;
     final MessageFrame frame =
         codeExecutor.executeCode(
@@ -70,8 +74,7 @@ public class ConstantinopleSStoreOperationGasCostTest {
             gasLimit,
             account -> account.setStorageValue(UInt256.ZERO, UInt256.valueOf(originalValue)));
     assertThat(frame.getState()).isEqualTo(State.COMPLETED_SUCCESS);
-    System.out.println(gasLimit - frame.getRemainingGas());
-    assertThat(frame.getRemainingGas()).isEqualTo(gasLimit - expectedGasUsed);
+    assertThat(frame.getRemainingGas()).isEqualTo(gasLimit - (expectedGasUsed + 2100));
     assertThat(frame.getGasRefund()).isEqualTo(expectedGasRefund);
   }