+++ evm/src/test/java/org/hyperledger/besu/evm/operation/MCopyOperationTest.java
@@ -12,15 +12,13 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.evm.operations;
+package org.hyperledger.besu.evm.operation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.hyperledger.besu.evm.EVM;
 import org.hyperledger.besu.evm.frame.MessageFrame;
 import org.hyperledger.besu.evm.gascalculator.CancunGasCalculator;
-import org.hyperledger.besu.evm.operation.MCopyOperation;
-import org.hyperledger.besu.evm.operation.Operation;
 import org.hyperledger.besu.evm.testutils.TestMessageFrameBuilder;
 
 import java.util.Arrays;